+++ lib/rubocop/cop/lint/binary_operator_with_identical_operands.rb
@@ -35,7 +35,7 @@ module RuboCop
           return unless node.binary_operation?
 
           lhs, operation, rhs = *node
-          return if MATH_OPERATORS.include?(node.method_name) && rhs.basic_literal?
+          return if MATH_OPERATORS.include?(node.method_name) && lhs.basic_literal?
 
           add_offense(node, message: format(MSG, op: operation)) if lhs == rhs
         end